+++ packages/botonic-plugin-flow-builder/src/action/first-interaction.ts
@@ -1,6 +1,6 @@
 import { FlowBuilderApi } from '../api'
-import { BOT_ACTION_PAYLOAD_PREFIX, MAIN_FLOW_NAME } from '../constants'
-import { FlowContent } from '../content-fields'
+import { MAIN_FLOW_NAME } from '../constants'
+import { FlowBotAction, FlowContent } from '../content-fields'
 import BotonicPluginFlowBuilder from '../index'
 import { getNodeByUserInput } from '../user-input'
 import { inputHasTextData } from '../utils'
@@ -16,6 +16,12 @@ export async function getContentsByFirstInteraction({
 }: FlowBuilderContext): Promise<FlowContent[]> {
   const firstInteractionContents =
     await flowBuilderPlugin.getStartContents(resolvedLocale)
+
+  // If the first interaction has a FlowBotAction, it should be the last content
+  // and avoid to render the match with keywords,intents or knowledge base
+  if (firstInteractionContents.at(-1) instanceof FlowBotAction) {
+    return firstInteractionContents
+  }
 
   if (inputHasTextData(request.input)) {
     const contentsByUserInput = await getContentsByUserInput({
@@ -49,12 +55,6 @@ async function getContentsByUserInput({
 
   if (request.input.payload === conversationStartId) {
     return []
-  }
-
-  if (request.input.payload?.startsWith(BOT_ACTION_PAYLOAD_PREFIX)) {
-    request.input.payload = flowBuilderPlugin.replaceBotActionPayload(
-      request.input.payload
-    )
   }
 
   const contentsByKeywordsOrIntents = await getContentsByPayload({