+++ packages/botonic-plugin-flow-builder/src/content-fields/hubtype-fields/bot-action.ts
@@ -1,9 +1,9 @@
 import { HtBaseNode } from './common'
-import { HtNodeWithoutContentType } from './node-types'
+import { HtNodeWithContentType } from './node-types'
 
 export interface HtBotActionNode extends HtBaseNode {
   id: string
-  type: HtNodeWithoutContentType.BOT_ACTION
+  type: HtNodeWithContentType.BOT_ACTION
   content: {
     payload_id: string
     payload_params?: string