+++ packages/botonic-plugin-flow-builder/tests/bot-action.test.ts
@@ -1,15 +1,12 @@
-import { INPUT } from '@botonic/core'
+import { BotonicAction, INPUT } from '@botonic/core'
 import { describe, test } from '@jest/globals'
 
-import { BOT_ACTION_PAYLOAD_PREFIX } from '../src/constants'
 import { FlowText } from '../src/index'
 import { ProcessEnvNodeEnvs } from '../src/types'
 import { basicFlow } from './helpers/flows/basic'
 import {
   createFlowBuilderPlugin,
   createFlowBuilderPluginAndGetContents,
-  createRequest,
-  getContentsAfterPreAndBotonicInit,
 } from './helpers/utils'
 
 describe('The user clicks on a button that is connected to a BotActionNode', () => {
@@ -17,7 +14,11 @@ describe('The user clicks on a button that is connected to a BotActionNode', () 
   const flowBuilderPlugin = createFlowBuilderPlugin({ flow: basicFlow })
   const ratingMessageUuid = '578b30eb-d230-4162-8a36-6c7fa18ff0db'
   const botActionUuid = '85dbeb56-81c9-419d-a235-4ebf491b4fc9'
-  test('The button has  a payload equal to ba|botActionUuid', async () => {
+  const ratingPayload = 'rating'
+  const payloadParms = '{"value":1,"followUpContentID":"SORRY"}'
+  const ratingPayloadWithParams = `${ratingPayload}|${payloadParms}`
+
+  test('The button has a payload equal to botActionUuid', async () => {
     const { contents } = await createFlowBuilderPluginAndGetContents({
       flowBuilderOptions: { flow: basicFlow },
       requestArgs: {
@@ -29,42 +30,29 @@ describe('The user clicks on a button that is connected to a BotActionNode', () 
     })
 
     const nextPaylod = (contents[0] as FlowText).buttons[0].payload
-    expect(nextPaylod).toBe(`${BOT_ACTION_PAYLOAD_PREFIX}${botActionUuid}`)
+    expect(nextPaylod).toBe(botActionUuid)
   })
 
-  test('The bot routes receive the correct payload', async () => {
-    const request = createRequest({
-      input: {
-        type: INPUT.POSTBACK,
-        payload: `${BOT_ACTION_PAYLOAD_PREFIX}${botActionUuid}`,
-      },
-      plugins: {
-        // @ts-ignore
-        flowBuilderPlugin,
+  test('The request.session._botonic_action have redirect:nextPayload', async () => {
+    const { contents, request } = await createFlowBuilderPluginAndGetContents({
+      flowBuilderOptions: { flow: basicFlow },
+      requestArgs: {
+        input: {
+          type: INPUT.POSTBACK,
+          payload: botActionUuid,
+        },
       },
     })
 
-    await flowBuilderPlugin.pre(request)
-    expect(request.input.payload).toBe(
-      'rating|{"value":1,"followUpContentID":"SORRY"}'
+    expect(contents.length).toBe(1)
+    expect(request.session._botonic_action).toBe(
+      `${BotonicAction.Redirect}:${ratingPayloadWithParams}`
     )
   })
 
   test('In the custom action the payloadParmas defined in the BotActionNode are obtained', async () => {
-    const request = createRequest({
-      input: {
-        type: INPUT.POSTBACK,
-        payload: `${BOT_ACTION_PAYLOAD_PREFIX}${botActionUuid}`,
-      },
-      plugins: {
-        // @ts-ignore
-        flowBuilderPlugin,
-      },
-    })
-
-    await flowBuilderPlugin.pre(request)
     const payloadParams = flowBuilderPlugin.getPayloadParams(
-      request.input.payload as string
+      ratingPayloadWithParams
     )
     expect(payloadParams).toEqual(
       JSON.parse('{"value":1,"followUpContentID":"SORRY"}')